requirements.txt
twilio
openai
fastapi
uvicorn
python-dotenv
